services/llm/pyproject.toml
[project]
name = "llm"
version = "0.1.0"
description = "LLM gateway service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "prometheus-client>=0.19",
    "shared",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm"]
